refactor: relocate physics engine to core engine layer
